Improve reliability of request cancellation tests
Added a short, one-second delay to each of the TestService's request methods to ensure the cancellation token is used before the http request is formed and sent. Also added Shopify's mock.shop API test endpoint as the TestService's target domain so an HttpRequest will actually send if it does start, rather than ending immediately with an invalid domain.