Change generic return type of `PostAsync<T>(GraphRequest)` to non-nullable `T`
The System.Text.Json deserializer will only return null in very limited edge-cases, chiefly when the json string is the json literal `null` or the deserialized type is a nullable primitive (e.g. `int?`, `bool?`) and the json structure doesn't match. Even an empty curly brace will not trip the json literal `null` case. However, the GraphService itself has multiple checks in place to ensure that the json string is not null, an empty curly brace or even a mismatched type – it expects `object` and will only operate on `object`, and it will throw an exception if it doesn't find and object. Given this, we can make the assumption that the deserialization will succeed and the object will not be null.