feat: add special handling for GraphQL request errors returned by Shopify
This commit adds support for parsing the top-level `errors` property returned by Shopify. If found, and the array isn't empty, the GraphService will throw a ShopifyGraphErrorsException which indicates there are issues with the request's mutation/query that the developer must fix. This exception is distinct from the ShopifyGraphUsersException, which indicates that there's something wrong with the operation that the end user may be able to fix.

(cherry picked from commit ea972c4ac888ef9ba3bdf8c3edd0da8f1d57ecc1)