fix: read request info before sending request
This change fixes a bug where ShopifySharp would throw an exception when running on .NET Framework when trying to read the request message's content after the request had already been sent. .NET Framework's HttpClient disposes the HttpRequestMessage's `Content` after sending the request (https://stackoverflow.com/a/25495500). When ShopifySharp tried to read the disposed request content, an `ObjectDisposedException` would be thrown. This bug did not occur outside of .NET Framework, as the request message's content is not disposed by the HttpClient there.

Closes #1080, #1129