converters: add GraphErrorExtensionsValueConverter to map string errors to dicts
Add a System.Text.Json converter which is responsible for mapping the
different json structures we've seen on the `GraphErrorExtensions.Value`
property. So far the property can have the following two structures:

1. `{"Value":{"foo":"bar","baz":"bat"}}`
2. `{"Value":"bar"}`

For backwards compatibility, the GraphErrorExtensionsValueConverter
takes structure #2 and maps it to structure #1, maintaining the
`IReadOnlyDictionary<string, object?>` type. An arbitrary key of
`$value` is assigned in this case.

Just speculation, but I believe that #2 happens when there is a problem
with the parameter passed to a GraphQL query/mutation; the string
`"bar"` here is the problematic parameter value. On the other hand, #1
represents a problem with a query or mutation structure, e.g. the data
being queried.

Fixes #1149