Deprecate BuildAuthorizationUrl string grant methods
This commit deprecates the `BuildAuthorizationUrl(..., IEnumerable<string>? grants =
null)` methods and provided guidance to use the
`BuildAuthorizationUrl(AuthorizationUrlOptions)` method instead. The goal is to
improve maintainability, as it's easier to add/remove/deprecate properties from a
record than it is to modify the parameters of a method without introducing breaking
changes.

Scope: oauth, deprecation