Begin reading from the pipe before the parser writes to it
Because it's designed for streaming, Pipe has a maximum buffer limit of around 4mb. If it isn't drained before reaching that limit, the writer will stall and wait for the reader to begin draining the pipe. The parser was waiting for the entire parsing operation to complete before it would start to read, but because Shopify's GraphQL document is so large, it would stall out before completing.

The parser now kicks off a task that begins reading the pipe's data before the parser begins parsing the GraphQL document, preventing the buffer from filling up and stalling out.

Scope: fix, parser