Add INode interface inheritance for any class or interface with an Id field
- Add appendINodeInheritedTypeIfAppropriate function to automatically inherit INode interface
- Apply INode inheritance to classes and interfaces that have Id fields but don't already inherit Node/INode
- Update writeClass and writeInterface functions to use the enhanced inheritance logic

While almost all of the types in Shopify's GraphQL schema use the `Node`
interface if they have an `Id` field, there are a handful that,
bizarrely, do not. This change overrides that behavior and forces any
class or interface which has an `Id` field to also inherit the `INode`
interface.