Replace polymorphic json attributes with a custom GraphUnionTypeConverter json converter
This change fixes a bug which prevented the deserialization of union types into their union cases due to the internal types used by ShopifySharp to hold the union case values (e.g. `CommentEventEmbedCustomer`, `CommentEventEmbedOrder`). The polymorphic attributes weren't able to determine how to deserialize into and out of those internal types (due to the `.Value` property on those internal types holding the actual value of the union case).

To fix this, the GraphUnionTypeConverter has been introduced and replaces the `JsonDerivedType` and `JsonPolymorphic` attributes. It handles deserialization and serialization of the union type, moving values into and out of the internal wrappers behind the scenes. It may even stand to reason that we don't need the internal wrappers at all with this converter – something to explore in the future.

Fixes #1213

type: fix
scope: generated-graphql