Add GraphQueryBuilder base class, implemented by generated types
Notes:

- The query builder argument method names do not need to be sanitized
against the list of C# and CLR keywords. Doing so leads to the argument
names turning into things like "AddArgument @metaspace".

- The `camelTypeName` must be using the type's name, not the pascal class name.
If we use the pascal class name, then name that gets serialized to a graph query
becomes suffixed with "QueryBuilder", so the querybuilder for every type would
think that the "foo" type was actually named "fooQueryBuilder"

type: feature
scope: querybuilders, sourcegen