Ensure union case uses explicit name parameter
Previously, union cases were added being added using the name passed to
their `Query` instead of the given `name` parameter. In other cases that
would be fine, but for adding unions to the selection we need to change
the name to `... on {typeName}`.

This fix uses the `AddField` callback to customize the union case's
name, passing in the explicit name parameter with the graphQL union case
syntax.

type: fix
scope: graphql, sourcegen